# 任务10/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "医学知识 RAG 问答服务",
    "description": "本地医学知识 RAG 问答服务。\n\n先从 PubMed oa_comm 语料检索证据，再由本地 `qwen3:8b` 依据证据作答，\n每个事实句都带 `[S#]` 出处，并在生成后逐条校验（阶段九强约束层）。\n\n* **统一响应体**：`{code, message, data, detail, request_id, timestamp, elapsed_ms}`，\n  成功与失败同一个形状；`code == 0` 为成功，码表见 `GET /api/v1/errors`。\n* **两种响应模式**：`POST /api/v1/qa/ask` 同步；`POST /api/v1/qa/stream` 走 SSE。\n* **一次问答 100 秒以上是常态**（四段链 + 层 D 校验），客户端超时请设足。\n",
    "version": "0.10.2"
  },
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "文档"
        ],
        "summary": "文献列表（过滤 + 游标分页）",
        "description": "按期刊 / 年份 / 标题关键词过滤，游标分页。\n\n· **没有 page 参数**：227 万行上的深翻页要扫掉前面所有行，所以只提供 `cursor`（传上一页返回的 `next_cursor`）。\n· `total` 默认不算（null），要就传 `with_total=true`：无过滤时约 24ms，配标题模糊搜最坏约 1.5s。\n· 标题搜是 `LIKE %kw%`，命中少时要全表扫，实测最坏 824ms。",
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "name": "journal",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "期刊全名，精确匹配",
              "title": "Journal"
            },
            "description": "期刊全名，精确匹配"
          },
          {
            "name": "pub_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1500
                },
                {
                  "type": "null"
                }
              ],
              "description": "发表年份，精确",
              "title": "Pub Year"
            },
            "description": "发表年份，精确"
          },
          {
            "name": "year_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1500
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year From"
            }
          },
          {
            "name": "year_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1500
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year To"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "标题包含该关键词（大小写不敏感）",
              "title": "Title"
            },
            "description": "标题包含该关键词（大小写不敏感）"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "上一页返回的 next_cursor",
              "title": "Cursor"
            },
            "description": "上一页返回的 next_cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "with_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "是否额外算总数（有代价，见上）",
              "default": false,
              "title": "With Total"
            },
            "description": "是否额外算总数（有代价，见上）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_DocumentPage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{doc_id}": {
      "get": {
        "tags": [
          "文档"
        ],
        "summary": "按 id 查文献",
        "description": "`doc_id` 传 PMCID（如 `PMC212698`）；本语料里 doc_id 与 pmcid 相同，两种都认。不在目录里 → **3001**（注意：那只表示不在这份 4M 抽样索引里，不代表 PMC 上没有这篇）。",
        "operationId": "get_document_api_v1_documents__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "description": "PMCID 或 doc_id",
              "title": "Doc Id"
            },
            "description": "PMCID 或 doc_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_DocumentIn_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DocumentIn": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id",
            "description": "文档 ID。本语料里等于 pmcid（distinct 只差 17）"
          },
          "pmcid": {
            "type": "string",
            "title": "Pmcid",
            "description": "PMC 号，唯一主键"
          },
          "pmid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pmid",
            "description": "PubMed 号；**不唯一**（勘误与原文共用），不能拿来当主键"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "文献标题（库内字段名 source_title）",
            "default": ""
          },
          "journal": {
            "type": "string",
            "title": "Journal",
            "default": ""
          },
          "pub_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pub Year",
            "description": "发表年份。任务书写的是 pub_date，但语料只保留到年份"
          },
          "abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract",
            "description": "摘要正文，由库内 section=abstract 的块拼回；**仅 7.7% 的文献有**，其余为 null"
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks",
            "description": "原文切块数（切块阶段记的，平均 28.6）",
            "default": 0
          },
          "indexed_chunks": {
            "type": "integer",
            "title": "Indexed Chunks",
            "description": "真正进了 4M 向量库的块数（平均 1.76）",
            "default": 0
          },
          "sections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sections",
            "description": "本篇在库内出现过的规范段落名"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "pmcid"
        ],
        "title": "DocumentIn",
        "description": "文献模型。**名字照任务书写的 `DocumentIn`**，不另起一个更\"像响应模型\"的名字，\n免得对照任务书时多一层翻译。\n\n与任务书字段清单的三处出入，都是语料决定的，不是省事：\n\n· **`pub_date` → `pub_year`**：切块阶段只保留了年份，库里 10 个元数据字段\n  （doc_id/chunk_index/total_chunks/source_title/token_count/section/pmcid/pmid/\n  journal/pub_year）里没有完整日期。放一个恒为 null 的 `pub_date` 只会让人以为\n  是数据缺失而去查。\n· **`abstract` 只有 7.7% 有值**（2,274,167 篇里 175,664 篇）。4M 索引是按**块**\n  抽样的，多数文献的摘要块根本没被抽中。字段保留（任务书点名了）但标可选，\n  填充率写在这里而不是等人自己发现。\n· **`total_chunks` 与 `indexed_chunks` 必须并列**：前者是原文切块数（平均 28.6），\n  后者是真正进了 4M 库的条数（平均 1.76）。只给前者，详情页会显示\"共 26 块\"\n  然后一块也列不出来。",
        "example": {
          "doc_id": "PMC212698",
          "indexed_chunks": 2,
          "journal": "PLoS Biology",
          "pmcid": "PMC212698",
          "pmid": "14551916",
          "pub_year": 2003,
          "sections": [
            "results",
            "methods"
          ],
          "title": "Candidate Gene Association Study in Type 2 Diabetes…",
          "total_chunks": 36
        }
      },
      "DocumentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentIn"
            },
            "type": "array",
            "title": "Items"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 20
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "下一页的 cursor；没有下一页时为 null"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total",
            "description": "只在 with_total=true 时计算，否则为 null"
          },
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters",
            "description": "本次真正生效的过滤条件"
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "DocumentPage",
        "description": "文献列表结果。**用游标不用页码**——`PageModel` 那套 `page/offset` 在 227 万行上\n会掉进翻页深渊（`LIMIT 20 OFFSET 100000` 得先扫掉前十万行）。游标是上一页最后一条\n的 pmcid，翻到第几页都是常数代价。\n\n`total` 默认为 null：无过滤时 `COUNT(*)` 实测 24ms、最坏（标题全表扫）1.5s，\n而翻页并不需要它。要就显式传 `with_total=true`。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseModel_DocumentIn_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务错误码，0 表示成功；码表见 GET /api/v1/errors",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "给人看的一句话说明",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "成功时的业务数据；失败时为 null"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "失败时的结构化补充信息"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的唯一 ID，与日志、SQLite 记录一一对应",
            "default": ""
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms",
            "description": "服务端处理耗时（毫秒）"
          }
        },
        "type": "object",
        "title": "ResponseModel[DocumentIn]",
        "example": {
          "code": 0,
          "data": {},
          "elapsed_ms": 12.3,
          "message": "ok",
          "request_id": "req-1a2b3c4d",
          "timestamp": "2026-08-10 12:00:00"
        }
      },
      "ResponseModel_DocumentPage_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务错误码，0 表示成功；码表见 GET /api/v1/errors",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "给人看的一句话说明",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentPage"
              },
              {
                "type": "null"
              }
            ],
            "description": "成功时的业务数据；失败时为 null"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "失败时的结构化补充信息"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的唯一 ID，与日志、SQLite 记录一一对应",
            "default": ""
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms",
            "description": "服务端处理耗时（毫秒）"
          }
        },
        "type": "object",
        "title": "ResponseModel[DocumentPage]",
        "example": {
          "code": 0,
          "data": {},
          "elapsed_ms": 12.3,
          "message": "ok",
          "request_id": "req-1a2b3c4d",
          "timestamp": "2026-08-10 12:00:00"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}